+++ web-wallet/src/routes/(app)/settings/__tests__/page.spec.js
@@ -9,6 +9,7 @@ import {
 } from "vitest";
 import { act, cleanup, fireEvent, render } from "@testing-library/svelte";
 import { get } from "svelte/store";
+import { setKey } from "lamb";
 
 import mockedWalletStore from "$lib/__mocks__/mockedWalletStore";
 import * as navigation from "$lib/navigation";
@@ -83,6 +84,18 @@ describe("Settings", () => {
     expect(container.firstChild).toMatchSnapshot();
   });
 
+  it("should show the wallet creation block height if it's greater than zero", () => {
+    settingsStore.update(setKey("walletCreationBlockHeight", 123n));
+
+    const { container, getByDisplayValue } = render(Settings, {});
+    const creationBlockInput = getByDisplayValue("123");
+
+    expect(creationBlockInput).toBeInTheDocument();
+    expect(container.firstChild).toMatchSnapshot();
+
+    settingsStore.reset();
+  });
+
   it("should disable the reset wallet button while a sync is in progress", async () => {
     const { getByRole } = render(Settings);
     const resetButton = getByRole("button", { name: /reset wallet/i });
@@ -146,7 +159,7 @@ describe("Settings", () => {
     const settingsResetSpy = vi.spyOn(settingsStore, "reset");
     const loginInfoStorageSpy = vi.spyOn(loginInfoStorage, "remove");
 
-    afterEach(() => {
+    beforeEach(() => {
       clearDataSpy.mockClear();
       confirmSpy.mockClear();
       settingsResetSpy.mockClear();