+++ web-wallet/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add support for partial unstake/claim rewards [#3009]
 - Add "Unstake" flow validation [#3009]
+- Store the wallet creation block height and show it in settings [#3381]
 
 ### Changed
 
@@ -547,6 +548,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#3354]: https://github.com/dusk-network/rusk/issues/3354
 [#3356]: https://github.com/dusk-network/rusk/issues/3356
 [#3362]: https://github.com/dusk-network/rusk/issues/3362
+[#3381]: https://github.com/dusk-network/rusk/issues/3381
 [#3387]: https://github.com/dusk-network/rusk/issues/3387
 
 <!-- VERSIONS -->